Ship a singleton of the Json serializer to reduce usage boilerplate
